Fix status bar updates
Before this change, we gave the command runner an outlet from the
TestStatusStatusBarView constructor. However, at the time that constructor is
run, the outlet hasn't been initialized yet and is always undefined.

This is most likely due to how Space Pen works internally.

With this change in place, we pass the outlet to the command runner every time
we want to run a command. This gives us a well defined outlet every time.